Extract PlanNodeWithCost
Extract PlanNodeWithCost to a top level class so that it can be used
with other cost-based optimization rules.
